CombatComputerAlpha
The first update to Weapons N' More. Contains the basic functions of the combat computer- predictive targeting. Sadly it does not work on AI controlled techs all too well at the moment, that will be fixed in a future update. But for player use the combat computer works just fine.